Bug Fixes
Add: size and empty functions to the device map

Fix: The argument checker would reject any table passed to it if the expected field wasn't a struct, even in the case of arrays.
Fix: Packets sent to the device sometimes received answers before the promise was added to the list, resulting in the reception to fail.
Fix: The orchestrator was simply removing the last 4 characters of a path when loading the user files. This caused issues when the path was a directory and therefore didn't have an extension.

Change: Log levels in some logs of the device