Add Error Generator
Added a menu where emergency messages can be generated. This is a
development tool designed to be used internally only, not by the end
user. It might still be neat to have for when users are able to develop
their own boards tho...

Fixed some minor warnings, playing around with R8L from the lua side of
things.